.history/requirements_20251013133232.txt
# --- Core Deep Learning Framework ---
torch
torchvision

# --- Hugging Face Ecosystem ---
# Untuk memuat dataset & menggunakan model dari Hub
datasets
transformers

# --- Model Library (Sangat Penting untuk ViT, Swin, ConvNeXt) ---
# Menyediakan implementasi SOTA yang mudah digunakan
timm

# --- Data Science & Numerics ---
numpy
pandas

# --- Image Processing & Visualization ---
Pillow
matplotlib
seaborn
opencv-python

# --- Machine Learning Utilities ---
# Untuk metrik seperti classification_report & confusion_matrix
scikit-learn

# --- Untuk Menjalankan Notebooks ---
jupyter
notebook